Revert CSS because class name same as attendance form
Also, the reason the form textarea is compressed is because we pulled in css from the attendance team that are using the same form-group class name with the property align-items: center. We should change the class name of the form so we can style it differently from them. Button class names also need to be changed.